refactor: extract resolver utilities to utils.ts (#25)
